requirements/dev.txt
-r base.txt

# Testing tools
factory_boy==2.6.0